Update back-end: 1. Menambahkan library numpy 2. Mengubah sedikit beberapa tipe data pada table database 3. Membuat form model 4. Menghapus view.py 5. Progress website sudah 85% 6. Form handling untuk perbandingan kriteria belum selesai 7. Mengubah sedikit style.css 8. Mengubah struktur file html agar data bisa ditampilkan dengan template Jinja2

+++ app/model/database_model.py
@@ -36,9 +36,9 @@ class Kriteria(SQLModel, table=True):
 class Perbandingan_Kriteria(SQLModel, table=True):
     id: int = Field(default=None, primary_key=True)
     id_kriteria1: int = Field(foreign_key="kriteria.id", ondelete="CASCADE")
-    nilai_kriteria1: float
+    nilai_kriteria1: int
     id_kriteria2: int = Field(foreign_key="kriteria.id", ondelete="CASCADE")
-    nilai_kriteria2: float
+    nilai_kriteria2: int
 
     # kriteria: Kriteria = Relationship(back_populates="perbandingan_kriteria")
     kriteria_1: Kriteria = Relationship(
